+++ build_runner/lib/src/asset/reader.dart
@@ -28,6 +28,12 @@ class AssetReaderSpy implements AssetReader {
   }
 
   @override
+  Future<List<int>> readAsBytes(AssetId id) {
+    _assetsRead.add(id);
+    return _delegate.readAsBytes(id);
+  }
+
+  @override
   Future<String> readAsString(AssetId input, {Encoding encoding: UTF8}) {
     _assetsRead.add(input);
     return _delegate.readAsString(input, encoding: encoding);