+++ build_runner/test/generate/serve_test.dart
@@ -11,19 +11,20 @@ import 'package:watcher/watcher.dart';
 
 import 'package:build_runner/build_runner.dart';
 import 'package:build_runner/src/server/server.dart' as server;
+import 'package:build_test/build_test.dart';
 
 import '../common/common.dart';
 
 void main() {
   group('serve', () {
-    InMemoryAssetWriter writer;
+    InMemoryRunnerAssetWriter writer;
     CopyBuilder copyBuilder;
     Phase copyAPhase;
     PhaseGroup copyAPhaseGroup;
 
     setUp(() {
       _terminateServeController = new StreamController();
-      writer = new InMemoryAssetWriter();
+      writer = new InMemoryRunnerAssetWriter();
 
       /// Basic phases/phase groups which get used in many tests
       copyBuilder = new CopyBuilder();
@@ -42,16 +43,16 @@ void main() {
       startServe(copyAPhaseGroup, {'a|web/a.txt': 'a'}, writer)
           .listen(results.add);
       var result = await nextResult(results);
-      checkOutputs({'a|web/a.txt.copy': 'a'}, result, writer.assets);
+      checkOutputs({'a|web/a.txt.copy': 'a'}, result.outputs, writer);
 
-      await writer.writeAsString(makeAsset('a|web/a.txt', 'b'));
+      await writer.writeAsString(makeAssetId('a|web/a.txt'), 'b');
       FakeWatcher.notifyWatchers(new WatchEvent(
           ChangeType.MODIFY, path.absolute('a', 'web', 'a.txt')));
 
       result = await nextResult(results);
       checkOutputs({
         'a|web/a.txt.copy': 'b',
-      }, result, writer.assets);
+      }, result.outputs, writer);
     });
 
     test('blocks serving files until the build is done', () async {
@@ -73,7 +74,7 @@ void main() {
       await wait(250);
       buildBlocker1.complete();
       var result = await nextResult(results);
-      checkOutputs({'a|web/a.txt.copy': 'a'}, result, writer.assets);
+      checkOutputs({'a|web/a.txt.copy': 'a'}, result.outputs, writer);
 
       /// Next request completes right away.
       var buildBlocker2 = new Completer();
@@ -84,7 +85,7 @@ void main() {
 
       /// Make an edit to force another build, and we should block again.
       copyBuilder.blockUntil = buildBlocker2.future;
-      await writer.writeAsString(makeAsset('a|web/a.txt', 'b'));
+      await writer.writeAsString(makeAssetId('a|web/a.txt'), 'b');
       FakeWatcher.notifyWatchers(new WatchEvent(
           ChangeType.MODIFY, path.absolute('a', 'web', 'a.txt')));
       // Give the build enough time to get started.
@@ -100,7 +101,7 @@ void main() {
       result = await nextResult(results);
       checkOutputs({
         'a|web/a.txt.copy': 'b',
-      }, result, writer.assets);
+      }, result.outputs, writer);
 
       /// Make sure we actually see the final request finish.
       return done.future;
@@ -112,13 +113,13 @@ final _debounceDelay = new Duration(milliseconds: 10);
 StreamController _terminateServeController;
 
 /// Start serving files and running builds.
-Stream<BuildResult> startServe(
-    PhaseGroup phases, Map<String, String> inputs, InMemoryAssetWriter writer) {
+Stream<BuildResult> startServe(PhaseGroup phases, Map<String, String> inputs,
+    InMemoryRunnerAssetWriter writer) {
   inputs.forEach((serializedId, contents) {
-    writer.writeAsString(makeAsset(serializedId, contents));
+    writer.writeAsString(makeAssetId(serializedId), contents);
   });
   final actualAssets = writer.assets;
-  final reader = new InMemoryAssetReader(actualAssets);
+  final reader = new InMemoryRunnerAssetReader(actualAssets);
   final rootPackage = new PackageNode('a', null, null, new Uri.file('a/'));
   final packageGraph = new PackageGraph.fromRoot(rootPackage);
   final watcherFactory = (String path) => new FakeWatcher(path);